requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
DateTime==5.5
discord.py==2.5.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
urllib3==2.3.0
yarl==1.18.3
zope.interface==7.2
